V0-Service Autor

+++ src/main/java/com/kata/books/repositories/AutorRepository.java
@@ -3,4 +3,4 @@ package com.kata.books.repositories;
 import com.kata.books.domain.Autor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AutorRepository extends JpaRepository<Autor, Long> {}
+public interface AutorRepository extends JpaRepository<Autor, Integer> {}
